rename defineMaxToken
