+++ frontend/app/chat/[chatId]/components/ChatInput/components/ConfigModal/ConfigModal.tsx
@@ -4,7 +4,7 @@ import { MdCheck, MdSettings } from "react-icons/md";
 import Button from "@/lib/components/ui/Button";
 import { Modal } from "@/lib/components/ui/Modal";
 import { models } from "@/lib/context/BrainConfigProvider/types";
-import { defineMaxTokens } from "@/lib/helpers/defineMexTokens";
+import { defineMaxTokens } from "@/lib/helpers/defineMaxTokens";
 
 import { useConfigModal } from "./hooks/useConfigModal";
