+++ frontend/lib/helpers/defineMaxTokens.ts
@@ -3,7 +3,7 @@ import { Model, PaidModels } from "../context/BrainConfigProvider/types";
 export const defineMaxTokens = (model: Model | PaidModels): number => {
   //At the moment is evaluating only models from OpenAI
   switch (model) {
-    case "gpt-3.5-turbo-0613":
+    case "gpt-3.5-turbo":
       return 500;
     case "gpt-3.5-turbo-16k":
       return 2000;
